[WFCORE-2771] Console should log bound ports
When specifying that Wildfly should bind the management console to an
ephemeral port (by specifying the port number as 0), Wildfly logs that
the management console is not enabled. In fact, this is not correct as
can be verified with lsof (or any other tool that enables seeing the
sockets that Wildfly has bound). The networking layer determined an
actual port number when binding the socket. Instead, Wildfly should log
the actual port that it is bound to. This commit causes this to be the
case.